Eu havia esquecido a condição que criei em alguns pontos (:

+++ backend/app/routers/test_delete_user.py
@@ -6,6 +6,11 @@ client = TestClient(app)
 def test_delete_user():
     user = {"name": "Ana"}
     create_response = client.post("/users/", json=user)
+    if create_response.status_code == 404:
+        print("Erro 404! Usuário não encontrado.")
+        print("Response content:", response.json())
+        return
+    
     user_id = create_response.json()["id"]
 
     response = client.delete(f"/users/{user_id}")
